# python-flask/swagger_server/models/cyclo.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Cyclo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, cyclo: str=None, name: str=None, active: bool=True, company: str=None):  # noqa: E501
        """Cyclo - a model defined in Swagger

        :param cyclo: The cyclo of this Cyclo.  # noqa: E501
        :type cyclo: str
        :param name: The name of this Cyclo.  # noqa: E501
        :type name: str
        :param active: The active of this Cyclo.  # noqa: E501
        :type active: bool
        :param company: The company of this Cyclo.  # noqa: E501
        :type company: str
        """
        self.swagger_types = {
            'cyclo': str,
            'name': str,
            'active': bool,
            'company': str
        }

        self.attribute_map = {
            'cyclo': 'cyclo',
            'name': 'name',
            'active': 'active',
            'company': 'company'
        }
        self._cyclo = cyclo
        self._name = name
        self._active = active
        self._company = company

    @classmethod
    def from_dict(cls, dikt) -> 'Cyclo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Cyclo of this Cyclo.  # noqa: E501
        :rtype: Cyclo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def cyclo(self) -> str:
        """Gets the cyclo of this Cyclo.

        Auto-generated primary key field  # noqa: E501

        :return: The cyclo of this Cyclo.
        :rtype: str
        """
        return self._cyclo

    @cyclo.setter
    def cyclo(self, cyclo: str):
        """Sets the cyclo of this Cyclo.

        Auto-generated primary key field  # noqa: E501

        :param cyclo: The cyclo of this Cyclo.
        :type cyclo: str
        """
        if cyclo is None:
            raise ValueError("Invalid value for `cyclo`, must not be `None`")  # noqa: E501

        self._cyclo = cyclo

    @property
    def name(self) -> str:
        """Gets the name of this Cyclo.


        :return: The name of this Cyclo.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Cyclo.


        :param name: The name of this Cyclo.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def active(self) -> bool:
        """Gets the active of this Cyclo.


        :return: The active of this Cyclo.
        :rtype: bool
        """
        return self._active

    @active.setter
    def active(self, active: bool):
        """Sets the active of this Cyclo.


        :param active: The active of this Cyclo.
        :type active: bool
        """

        self._active = active

    @property
    def company(self) -> str:
        """Gets the company of this Cyclo.

        This property is a reference to a Company  # noqa: E501

        :return: The company of this Cyclo.
        :rtype: str
        """
        return self._company

    @company.setter
    def company(self, company: str):
        """Sets the company of this Cyclo.

        This property is a reference to a Company  # noqa: E501

        :param company: The company of this Cyclo.
        :type company: str
        """
        if company is None:
            raise ValueError("Invalid value for `company`, must not be `None`")  # noqa: E501

        self._company = company