# python-flask/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: bampli-api
  description: The API for the Business Amplifier project
  contact:
    name: Jose Motta
    url: http://bampli.com
    email: josemotta@bampli.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
servers:
- url: https://virtserver.swaggerhub.com/motta/bampli/1.0.0-oas3
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/motta/bampli/1.0.0
...
paths:
...
  /cyclo/:
    get:
      tags:
      - Cyclo
      summary: Load the list of Cyclos
      description: Loads list of Cyclos
      operationId: search_cyclos
      parameters:
      - name: size
        in: query
        description: Size of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: integer
        x-example: 10
      - name: page
        in: query
        description: Number of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: number
        x-example: 1
      - name: sort
        in: query
        description: Order in which to retrieve the results.
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: name ASC
      - name: name
        in: query
        description: Allows to filter the collections of result by name
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: George Street Brewery
      - name: company
        in: query
        description: Allows to filter the collections of result by company
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: 0e8cedd0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          headers:
            X-Page-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Number:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Size:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 25
            X-Total-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 2
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cyclo'
                x-content-type: application/json
              examples:
                response:
                  value: |-
                    [{
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "P&Q Production Line",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }]
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    post:
      tags:
      - Cyclo
      summary: Create a new Cyclo
      description: Adds a Cyclo
      operationId: add_cyclo
      requestBody:
        $ref: '#/components/requestBodies/Cyclo'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cyclo'
              examples:
                response:
                  value: |-
                    {
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "P&Q Production Line",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
  /cyclo/{cycloid}:
    get:
      tags:
      - Cyclo
      summary: Load an individual Cyclo
      description: Loads a Cyclo
      operationId: get_cyclo
      parameters:
      - name: cycloid
        in: path
        description: Identifier of the Cyclo
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cyclo'
              examples:
                response:
                  value: |-
                    {
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "P&Q Production Line",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    put:
      tags:
      - Cyclo
      summary: Update a Cyclo
      description: Stores a Cyclo
      operationId: update_cyclo
      parameters:
      - name: cycloid
        in: path
        description: Identifier of the Cyclo
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      requestBody:
        $ref: '#/components/requestBodies/Cyclo'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cyclo'
              examples:
                response:
                  value: |-
                    {
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "P&Q Production Line",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    delete:
      tags:
      - Cyclo
      summary: Delete a Cyclo
      description: Deletes a Cyclo
      operationId: delete_cyclo
      parameters:
      - name: cycloid
        in: path
        description: Identifier of the Cyclo
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
  /stage/:
    get:
      tags:
      - Cyclo
      summary: Load the list of Stages
      description: Loads list of Stages
      operationId: search_stages
      parameters:
      - name: size
        in: query
        description: Size of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: integer
        x-example: 10
      - name: page
        in: query
        description: Number of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: number
        x-example: 1
      - name: sort
        in: query
        description: 'Order in which to retrieve the results. Multiple sort criteria
          can be passed. Example: sort=name ASC'
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: name ASC
      - name: name
        in: query
        description: Allows to filter the collections of result by name
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: Painting
      - name: cyclo
        in: query
        description: Allows to filter the collections of result by cyclo
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: 0e8cedd0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          headers:
            X-Page-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Number:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Size:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 25
            X-Total-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 2
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stage'
                x-content-type: application/json
              examples:
                response:
                  value: |-
                    [{
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Painting",
                      "active": true,
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }]
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    post:
      tags:
      - Cyclo
      summary: Create a new Stage
      description: Adds a Stage
      operationId: add_stage
      requestBody:
        $ref: '#/components/requestBodies/Stage'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stage'
              examples:
                response:
                  value: |-
                    {
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Painting",
                      "active": true,
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
  /stage/{stageid}:
    get:
      tags:
      - Cyclo
      summary: Load an individual Stage
      description: Loads a Stage
      operationId: get_stage
      parameters:
      - name: stageid
        in: path
        description: Identifier of the Stage
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stage'
              examples:
                response:
                  value: |-
                    {
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Painting",
                      "active": true,
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    put:
      tags:
      - Cyclo
      summary: Update a Stage
      description: Stores a Stage
      operationId: update_stage
      parameters:
      - name: stageid
        in: path
        description: Identifier of the Stage
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      requestBody:
        $ref: '#/components/requestBodies/Stage'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stage'
              examples:
                response:
                  value: |-
                    {
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Painting",
                      "active": true,
                      "cyclo": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    delete:
      tags:
      - Cyclo
      summary: Delete a Stage
      description: Deletes a Stage
      operationId: delete_stage
      parameters:
      - name: stageid
        in: path
        description: Identifier of the Stage
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
  /task/:
    get:
      tags:
      - Cyclo
      summary: Load the list of Tasks
      description: Loads list of Tasks
      operationId: search_tasks
      parameters:
      - name: size
        in: query
        description: Size of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: integer
        x-example: 10
      - name: page
        in: query
        description: Number of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: number
        x-example: 1
      - name: sort
        in: query
        description: 'Order in which to retrieve the results. Multiple sort criteria
          can be passed. Example: sort=name ASC,city DESC'
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: name ASC
      - name: name
        in: query
        description: Allows to filter the collections of result by name
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: George Street Brewery
      - name: stage
        in: query
        description: Allows to filter the collections of result by stage
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: 0e8cedd0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          headers:
            X-Page-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Number:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Size:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 25
            X-Total-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 2
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                x-content-type: application/json
              examples:
                response:
                  value: |-
                    [{
                      "task": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Paint",
                      "active": true,
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }]
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    post:
      tags:
      - Cyclo
      summary: Create a new Task
      description: Adds a Task
      operationId: add_task
      requestBody:
        $ref: '#/components/requestBodies/Task'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              examples:
                response:
                  value: |-
                    {
                      "task": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Paint",
                      "active": true,
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
  /task/{taskid}:
    get:
      tags:
      - Cyclo
      summary: Load an individual Task
      description: Loads a Task
      operationId: get_task
      parameters:
      - name: taskid
        in: path
        description: Identifier of the Task
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              examples:
                response:
                  value: |-
                    {
                      "task": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Paint",
                      "active": true,
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    put:
      tags:
      - Cyclo
      summary: Update a Task
      description: Stores a Task
      operationId: update_task
      parameters:
      - name: taskid
        in: path
        description: Identifier of the Task
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      requestBody:
        $ref: '#/components/requestBodies/Task'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              examples:
                response:
                  value: |-
                    {
                      "task": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Paint",
                      "active": true,
                      "stage": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
    delete:
      tags:
      - Cyclo
      summary: Delete a Task
      description: Deletes a Task
      operationId: delete_task
      parameters:
      - name: taskid
        in: path
        description: Identifier of the Task
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.cyclo_controller
...
components:
  schemas:
    Error:
      required:
      - code
      type: object
      properties:
        code:
          maximum: 599
          minimum: 400
          type: integer
        description:
          type: string
          example: 'Bad query parameter [$size]: Invalid integer value [abc]'
        reasonPhrase:
          type: string
          example: Bad Request
      description: This general error structure is used throughout this API.
      example: |-
        {
          "code": 400,
          "description": "Bad query parameter [$size]: Invalid integer value [abc]",
          "reasonPhrase": "Bad Request"
        }
...
    Cyclo:
      required:
      - company
      - cyclo
      - name
      type: object
      properties:
        cyclo:
          type: string
          description: Auto-generated primary key field
          example: 0e8ffb10-ad98-11e6-bf2e-47644ada7c0f
        name:
          type: string
          example: P&Q Production Line
        active:
          type: boolean
          default: true
        company:
          type: string
          description: This property is a reference to a Company
          example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      description: A representation of a Cyclo.
      example: |-
        {
          "cyclo": "0e8dd830-ad98-11e6-bf2e-47644ada7c0f",
          "name": "P&Q Production Line",
          "active": true,
          "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f"
        }
    Stage:
      required:
      - cyclo
      - name
      - stage
      type: object
      properties:
        stage:
          type: string
          description: Auto-generated primary key field
          example: 0e8ffb10-ad98-11e6-bf2e-47644ada7c0f
        name:
          type: string
          example: Assembly Station
        active:
          type: boolean
          default: true
        cyclo:
          type: string
          description: This property is a reference to a Cyclo
          example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      description: A representation of a Stage.
      example: |-
        {
          "stage": "0e8dd830-ad98-11e6-bf2e-47644ada7c0f",
          "name": "Assembly Station",
          "active": true,
          "cyclo": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f"
        }
    Task:
      required:
      - name
      - stage
      - task
      type: object
      properties:
        task:
          type: string
          description: Auto-generated primary key field
          example: 0e8ffb10-ad98-11e6-bf2e-47644ada7c0f
        name:
          type: string
          example: Painting
        active:
          type: boolean
          default: true
        stage:
          type: string
          description: This property is a reference to a Stage
          example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      description: A representation of a Task.
      example: |-
        {
          "task": "0e8dd830-ad98-11e6-bf2e-47644ada7c0f",
          "name": "Painting",
          "active": true,
          "stage": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f"
        }
...
  requestBodies:
    Task:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Task'
      required: true
    Stage:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Stage'
      required: true
    Cyclo:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Cyclo'
      required: true
...
  securitySchemes:
    HTTP_BASIC:
      type: http
      description: All GET methods are public, meaning that *you can read all the
        data*. Write operations require authentication and therefore are forbidden
        to the general public.
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_HTTP_BASIC